% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klein-vella-data.R
\name{create_klein_vella_config}
\alias{create_klein_vella_config}
\title{Create configuration for Klein & Vella data generation}
\usage{
create_klein_vella_config(
  n,
  k = 1,
  beta1,
  beta2,
  gamma1,
  rho,
  delta1,
  delta2,
  x_dist = "normal",
  seed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{n}{Sample size}

\item{k}{Number of exogenous variables (default = 1)}

\item{beta1}{Coefficients for Y1 equation (length k+1 including intercept)}

\item{beta2}{Coefficients for Y2 equation (length k+1 including intercept)}

\item{gamma1}{Endogenous parameter in Y1 equation}

\item{rho}{Correlation between structural errors}

\item{delta1}{Variance function parameters for epsilon1 (length k+1)}

\item{delta2}{Variance function parameters for epsilon2 (length k+1)}

\item{x_dist}{Distribution of X variables ("normal", "uniform", "mixed")}

\item{seed}{Random seed}

\item{verbose}{Whether to print information}
}
\value{
List containing configuration parameters
}
\description{
Create configuration for Klein & Vella data generation
}
\examples{
# Simple example with one X variable
config <- create_klein_vella_config(
  n = 1000,
  beta1 = c(0.5, 1.5),
  beta2 = c(1.0, -1.0),
  gamma1 = -0.8,
  rho = 0.6,
  delta1 = c(0.1, 0.3),
  delta2 = c(0.2, -0.2)
)

}
