% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-constants.R
\docType{data}
\name{.ALPHA_DEFAULT}
\alias{.ALPHA_DEFAULT}
\title{Internal numeric and string constants for the hetid package}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
.ALPHA_DEFAULT
}
\description{
This file defines package-wide constants to ensure Single Source of Truth.
It is loaded first due to alphabetical ordering (00- prefix).
}
\keyword{internal}
