% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{.hetid_constants}
\alias{.hetid_constants}
\title{Internal constants for the hetid package}
\format{
An object of class \code{environment} of length 18.
}
\usage{
.hetid_constants
}
\description{
This file defines all package-wide constants using a single
environment to ensure Single Source of Truth and follow R package best practices.
}
\keyword{internal}
