% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-hetid.R
\name{ensure_stata_packages}
\alias{ensure_stata_packages}
\title{Helper to ensure Stata packages are installed}
\usage{
ensure_stata_packages()
}
\description{
Helper to ensure Stata packages are installed
}
