% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prono-diagonal-garch.R
\name{estimate_dynamic_correlation}
\alias{estimate_dynamic_correlation}
\title{Estimate Dynamic Correlation}
\usage{
estimate_dynamic_correlation(z1, z2, lambda = 0.94)
}
\arguments{
\item{z1}{Standardized residuals series 1}

\item{z2}{Standardized residuals series 2}

\item{lambda}{Smoothing parameter (0.94 typical)}
}
\value{
Vector of time-varying correlations
}
\description{
Simple exponential smoothing for correlation dynamics
}
\keyword{internal}
