% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klein-vella-data.R
\name{generate_klein_vella_data}
\alias{generate_klein_vella_data}
\title{Generate data following Klein & Vella (2010) assumptions}
\usage{
generate_klein_vella_data(config, return_true_values = FALSE)
}
\arguments{
\item{config}{Configuration object from create_klein_vella_config or a list}

\item{return_true_values}{Whether to return true variance functions}
}
\value{
Data frame with generated data
}
\description{
Generate data following Klein & Vella (2010) assumptions
}
\details{
Generates data from the triangular system:
Y1 = X'beta1 + gamma1*Y2 + epsilon1
Y2 = X'beta2 + epsilon2

where the errors have:
\itemize{
\item Constant conditional correlation: Corr(epsilon1, epsilon2 | X) = rho
\item Heteroskedastic variances: Var(epsilon_j | X) = exp(X'delta_j)
}
}
\examples{
config <- create_klein_vella_config(
  n = 1000,
  beta1 = c(0.5, 1.5),
  beta2 = c(1.0, -1.0),
  gamma1 = -0.8,
  rho = 0.6,
  delta1 = c(0.1, 0.3),
  delta2 = c(0.2, -0.2)
)
data <- generate_klein_vella_data(config)
head(data)

}
