% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-df.R
\name{get_critical_value}
\alias{get_critical_value}
\title{Get critical value for confidence intervals}
\usage{
get_critical_value(n, k, alpha = .ALPHA_DEFAULT, df_adjust = "asymptotic")
}
\arguments{
\item{n}{Sample size}

\item{k}{Number of parameters}

\item{alpha}{Significance level (default 0.05)}

\item{df_adjust}{Character string: "asymptotic" (default) or "finite"}
}
\value{
Critical value
}
\description{
Get critical value for confidence intervals
}
