% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{hetid_const}
\alias{hetid_const}
\title{Get package constants}
\usage{
hetid_const(constant_name)
}
\arguments{
\item{constant_name}{Character. Name of the constant to retrieve}
}
\value{
The constant value from the package's constants environment
}
\description{
Get package constants
}
\keyword{internal}
