% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klein-vella-estimation.R
\name{klein_vella_parametric}
\alias{klein_vella_parametric}
\title{Parametric Klein & Vella estimation}
\usage{
klein_vella_parametric(
  data,
  y1_var = "Y1",
  y2_var = "Y2",
  x_vars = NULL,
  variance_type = "exponential",
  optimization_method = "BFGS",
  verbose = TRUE
)
}
\arguments{
\item{data}{Data frame containing Y1, Y2, and X variables}

\item{y1_var}{Name of first endogenous variable (default: "Y1")}

\item{y2_var}{Name of second endogenous variable (default: "Y2")}

\item{x_vars}{Names of exogenous variables (default: auto-detect)}

\item{variance_type}{Type of variance function ("exponential", "power", "linear")}

\item{optimization_method}{Optimization method for nonlinear least squares}

\item{verbose}{Whether to print progress information}
}
\value{
List containing estimation results
}
\description{
Parametric Klein & Vella estimation
}
\details{
Implements the parametric version of Klein & Vella (2010) using the
control function approach with parametric variance specifications.

The model is:
Y1 = X'beta1 + gamma1\emph{Y2 + rho}(S1(X)/S2(X))*epsilon2 + eta1

where S1(X) and S2(X) are conditional standard deviations.
}
\examples{
# Generate data
config <- create_klein_vella_config(
  n = 1000,
  beta1 = c(0.5, 1.5),
  beta2 = c(1.0, -1.0),
  gamma1 = -0.8,
  rho = 0.6,
  delta1 = c(0.1, 0.3),
  delta2 = c(0.2, -0.2)
)
data <- generate_klein_vella_data(config)

# Estimate
results <- klein_vella_parametric(data)
print(results)

}
