% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lewbel-gmm.R
\name{lewbel-gmm}
\alias{lewbel-gmm}
\title{GMM Estimation for Lewbel (2012) Heteroskedasticity-Based Identification}
\description{
Implementation of Generalized Method of Moments (GMM) estimation
for Lewbel's (2012) heteroskedasticity-based identification strategy.
}
\keyword{internal}
