% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klein-vella-estimation.R
\name{print.klein_vella_fit}
\alias{print.klein_vella_fit}
\title{Print method for Klein & Vella estimation results}
\usage{
\method{print}{klein_vella_fit}(x, ...)
}
\arguments{
\item{x}{A klein_vella_fit object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for Klein & Vella estimation results
}
