% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klein-vella-semiparametric.R
\name{print.klein_vella_semipar}
\alias{print.klein_vella_semipar}
\title{Print method for semiparametric Klein & Vella results}
\usage{
\method{print}{klein_vella_semipar}(x, ...)
}
\arguments{
\item{x}{A klein_vella_semipar object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for semiparametric Klein & Vella results
}
