% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lewbel-gmm.R
\name{print.lewbel_gmm}
\alias{print.lewbel_gmm}
\title{Print Method for Lewbel GMM}
\usage{
\method{print}{lewbel_gmm}(x, ...)
}
\arguments{
\item{x}{An object of class "lewbel_gmm".}

\item{...}{Additional arguments passed to print.gmm.}
}
\description{
Print Method for Lewbel GMM
}
