% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klein-vella-estimation.R
\name{print.summary.klein_vella_fit}
\alias{print.summary.klein_vella_fit}
\title{Print summary of Klein & Vella estimation results}
\usage{
\method{print}{summary.klein_vella_fit}(x, ...)
}
\arguments{
\item{x}{A summary.klein_vella_fit object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print summary of Klein & Vella estimation results
}
