% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klein-vella-estimation.R
\name{run_klein_vella_demo}
\alias{run_klein_vella_demo}
\title{Run Klein & Vella demonstration}
\usage{
run_klein_vella_demo(n = .hetid_const("N_SMALL"), verbose = TRUE)
}
\arguments{
\item{n}{Sample size (default: 500)}

\item{verbose}{Whether to print results}
}
\value{
Invisibly returns the results
}
\description{
Run Klein & Vella demonstration
}
\examples{
run_klein_vella_demo()

}
