% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lewbel-monte-carlo.R
\name{run_lewbel_demo}
\alias{run_lewbel_demo}
\title{Run Quick Lewbel Monte Carlo Demo}
\usage{
run_lewbel_demo(num_simulations = 100, verbose = TRUE)
}
\arguments{
\item{num_simulations}{Integer. Number of simulations to run (default: 100).}

\item{verbose}{Logical. Whether to print progress messages (default: TRUE).}
}
\value{
Results from run_lewbel_monte_carlo() with reduced parameters.
}
\description{
Runs a quick demonstration of the Lewbel Monte Carlo simulation with
reduced parameters for faster execution.
}
\examples{
\dontrun{
# Quick demo with 50 simulations
demo_results <- run_lewbel_demo(50)

# Silent demo
demo_results <- run_lewbel_demo(100, verbose = FALSE)
}

}
\references{
Lewbel, A. (2012). Using heteroscedasticity to identify and estimate
mismeasured and endogenous regressor models. Journal of Business & Economic
Statistics, 30(1), 67-80. \doi{10.1080/07350015.2012.643126}
}
\seealso{
\code{\link{run_lewbel_monte_carlo}}
}
