% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prono.R
\name{run_prono_demo}
\alias{run_prono_demo}
\title{Run Prono demonstration}
\usage{
run_prono_demo(n = 500, print_results = TRUE)
}
\arguments{
\item{n}{Sample size}

\item{print_results}{Whether to print results}
}
\value{
Results from single simulation
}
\description{
Run Prono demonstration
}
