% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prono.R
\name{run_prono_monte_carlo}
\alias{run_prono_monte_carlo}
\title{Run Prono Monte Carlo simulation}
\usage{
run_prono_monte_carlo(
  config,
  n_sims = 1000,
  parallel = FALSE,
  n_cores = NULL,
  progress = TRUE
)
}
\arguments{
\item{config}{Configuration list}

\item{n_sims}{Number of simulations}

\item{parallel}{Whether to use parallel processing}

\item{n_cores}{Number of cores for parallel processing}

\item{progress}{Whether to show progress bar}
}
\value{
Data frame with simulation results
}
\description{
Run Prono Monte Carlo simulation
}
