% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klein-vella-estimation.R
\name{summary.klein_vella_fit}
\alias{summary.klein_vella_fit}
\title{Summary method for Klein & Vella estimation results}
\usage{
\method{summary}{klein_vella_fit}(object, ...)
}
\arguments{
\item{object}{A klein_vella_fit object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary method for Klein & Vella estimation results
}
