% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lewbel-gmm.R
\name{summary.lewbel_gmm}
\alias{summary.lewbel_gmm}
\title{Summary Method for Lewbel GMM}
\usage{
\method{summary}{lewbel_gmm}(object, ...)
}
\arguments{
\item{object}{An object of class "lewbel_gmm".}

\item{...}{Additional arguments passed to summary.gmm.}
}
\description{
Summary Method for Lewbel GMM
}
