% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-hetid.R
\name{generate_hetid_test_data}
\alias{generate_hetid_test_data}
\title{Generate consistent test data for all comparisons}
\usage{
generate_hetid_test_data(n = 1000, seed = 42)
}
\description{
Generate consistent test data for all comparisons
}
